# Conflicts:
#	Desarrollo/BulletTest/.vs/Bu
lletTest/v14/.suo

+++ Desarrollo/BulletTest/BulletTest/Entities/PhysicsEntity.cpp
@@ -27,7 +27,7 @@ void PhysicsEntity::rotate(Vec3<float> rot)
 {
 
 	btQuaternion quat;
-	//quat.setEulerZYX(rot.getZ(), rot.getY(), rot.getX()); //or quat.setEulerZYX depending on the ordering you want
+	//or quat.setEulerZYX depending on the ordering you want
 	quat.setEulerZYX(rot.getZ(), rot.getY(), rot.getX());
 	m_rigidBody->getWorldTransform().setRotation(quat);
 
@@ -38,26 +38,27 @@ void PhysicsEntity::inicializar()
 	update(Time::Zero);
 }
 
+//Metodo static para poder llamarlo desde otros sitios, es mas bien un metodo para llevarlo a Util.h
 static Vec3<float> toEulerianAngle(const btQuaternion& q)
 {
 	Vec3<float> result;
 
-	double ysqr = q.y() * q.y();
+	float ysqr = q.y() * q.y();
 
 	// roll (x-axis rotation)
-	double t0 = +2.0 * (q.w() * q.x() + q.y() * q.z());
-	double t1 = +1.0 - 2.0 * (q.x() * q.x() + ysqr);
+	float t0 = +2.0 * (q.w() * q.x() + q.y() * q.z());
+	float t1 = +1.0 - 2.0 * (q.x() * q.x() + ysqr);
 	result.setX(std::atan2(t0, t1));
 
 	// pitch (y-axis rotation)
-	double t2 = +2.0 * (q.w() * q.y() - q.z() * q.x());
+	float t2 = +2.0 * (q.w() * q.y() - q.z() * q.x());
 	t2 = t2 > 1.0 ? 1.0 : t2;
 	t2 = t2 < -1.0 ? -1.0 : t2;
 	result.setY(std::asin(t2));
 
 	// yaw (z-axis rotation)
-	double t3 = +2.0 * (q.w() * q.z() + q.x() * q.y());
-	double t4 = +1.0 - 2.0 * (ysqr + q.z() * q.z());
+	float t3 = +2.0 * (q.w() * q.z() + q.x() * q.y());
+	float t4 = +1.0 - 2.0 * (ysqr + q.z() * q.z());
 
 	result.setZ(std::atan2(t3, t4));
 
@@ -69,31 +70,22 @@ void PhysicsEntity::update(Time elapsedTime)
 
 
 	// Set rotation
-	//btVector3 Euler;
 	const btQuaternion& TQuat = m_rigidBody->getOrientation();
-	//quaternion q(TQuat.getX(), TQuat.getY(), TQuat.getZ(), TQuat.getW());
-	//q.toEuler(cons(Euler));
 
+	//Pasamos el quaternion a angulos euler
 	Vec3<float> Euler = toEulerianAngle(TQuat);
 
 	Euler *= RADTODEG;
 
-	//m_nodo->setRotationXYZ(Vec3<float>(-m_renderRotation.getX(), m_renderRotation.getY() + 180, -m_renderRotation.getZ()));
 	m_rotation = Vec3<float>(-Euler.getX(), Euler.getY() + 180, -Euler.getZ());
 	m_nodo->setRotationXYZ(m_rotation);
-	//m_renderState.updateRotations();
 
-	// Set position
-	//m_rigidBody->setm
-	//btVector3 Point2 = m_rigidBody->getCenterOfMassTransform().getOrigin();
+
+	// Set position restando el center collision para centrar la colision al centro del objeto
 	btVector3 Point = m_rigidBody->getCenterOfMassPosition();
-	//if(centerCollision==NULL)
 
 	m_position = Vec3<float>((float)Point[0] - centerCollision.getX(), (float)Point[1] - centerCollision.getY(), (float)Point[2] - centerCollision.getX());
 	m_nodo->setPosition(m_position);
-	//m_renderState.updatePositions();
-
-	//m_rigidBody->pos
 }
 
 void PhysicsEntity::handleInput()
